fixed except spotify stream
