+++ styles.css
@@ -50,6 +50,14 @@ html {
   color: #e2c044;
 }
 
+.hidden {
+  display: none !important;
+}
+
+.img-error {
+  visibility: hidden;
+}
+
 /* Content Area */
 #content {
   display: flex;
@@ -336,13 +344,15 @@ footer {
   background: #d4b13e;
 }
 @media (max-width: 900px) {
-  #library .items {
+  #library .items,
+  #explore .items.showing-tracklist {
     display: flex !important;
     flex-direction: column !important;
   }
 
   /* Album info always visible at top - no scrolling */
-  .library-grid.hide-grid+.library-tracklist+.library-album-info {
+  .library-grid.hide-grid+.library-tracklist+.library-album-info,
+  #explore .items.showing-tracklist .library-album-info {
     width: 100% !important;
     height: auto !important;
     max-height: none !important;
@@ -363,7 +373,8 @@ footer {
   }
 
   /* Tracklist handles all scrolling */
-  .library-grid.hide-grid+.library-tracklist {
+  .library-grid.hide-grid+.library-tracklist,
+  #explore .items.showing-tracklist .library-tracklist {
     width: 100% !important;
     max-height: calc(100vh - 300px) !important;
     position: static !important;
@@ -819,13 +830,28 @@ footer {
 }
 
 /* Explore Tab */
-#explore .items {
+.explore-browse {
   display: flex;
   flex-direction: column;
   padding: 20px;
-  width: calc(100% - 138px);
+  width: 100%;
   height: 100%;
   box-sizing: border-box;
+  overflow: hidden;
+}
+
+#explore .items {
+  position: relative;
+  width: calc(100% - 138px);
+  height: 100%;
+  overflow: hidden;
+  pointer-events: auto;
+  align-items: center;
+}
+
+#explore .items.showing-tracklist {
+  display: flex;
+  flex-direction: row;
   overflow: hidden;
 }
 
@@ -921,35 +947,6 @@ footer {
   margin-bottom: 20px;
 }
 
-.explore-tracklist {
-  flex: 55;
-  overflow-y: auto;
-  border-right: 2px solid rgba(255, 255, 255, 0.1);
-  padding-right: 20px;
-}
-
-.explore-album-info {
-  flex: 45;
-  padding-left: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.explore-back-btn {
-  color: #e2c044;
-  font-family: oswald;
-  font-size: 16px;
-  cursor: pointer;
-  margin-bottom: 15px;
-  display: inline-block;
-  transition: transform 0.2s;
-}
-
-.explore-back-btn:hover {
-  transform: scale(1.1);
-}
-
 /* Spotify Auth */
 #spotifyAuth {
   display: flex;
@@ -957,6 +954,9 @@ footer {
   gap: 12px;
   width: 80%;
   max-width: 800px;
+  margin-top: 40px;
+  border-top: 2px solid #483666;
+  padding-top: 30px;
 }
 
 #spotifyAuth label {
